fix: keep background setup files under the max-lines ratchet
Split oversized workspace-create sources, refresh Riverpod hashes, and drop unused mobile result bindings so PR checks can pass.
