+++ mobile/lib/src/features/workbench/presentation/create_workspace_prompt.dart
@@ -307,7 +307,7 @@ extension _CreateWorkspacePromptForm on _CreateWorkspaceScreenState {
         );
     try {
       if (_createAnother) {
-        final result = await future;
+        await future;
         if (!mounted) {
           return;
         }
